Changed beaver sprite to animated .gif
Beaver is now currently displayed as Beaver_Walking.gif, using the
ImageIcon class (javax.swing.ImageIcon).
In the future, the images associated to an entity will be dependant on
state and will be stored with the entity information. This way each
entity can simply be told to draw, and the correct image/animation will
be drawn for that entity in its current state.