Added ResourceHandling package, with ResourceCollection and ResourceState classes.
ResourceState will be used in ResourceCollection. Entity classes will
use a single ResourceCollection to associate assets (sprites and sounds)
with any given state for such entity. As well as information on how
(where and when) to use them, in respect to hitbox (for visual assets),
and time/state-change (for audio).