Cleaned up entity classes. Removed type-specific classes in favour of using CombatEntity for all NPCs.
Removed type-specific classes in favour of using CombatEntity for all
NPCs, providing them with unique model resources and behaviour patterns.
This will allow for more modular and modifiable entities.
eg. it wouldn't be difficult to spawn a Moose with a Bear's stats and
LumberJack behaviour pattern (in case you ever wanted to).