Changed Behaviour class/interface structure.
Switched the Behaviour interface with BehaviourInterface, making
Behaviour a class.
Entity-specific behaviour classes will implement the BehaviourInterface.
Behaviour will then call to the static methods in the specific behaviour
classes corresponding with the type of entity associated to the call.
This will allow each aspect of an entity to be defined separately:
visual assets, audio assets, behaviour, combat stats, etc.