Overhauled State/StateInterface, Entity and ResourceCollection
ResourceCollection now handles the association of assets to entity
states. Given the (String) name of a state, it can return the assets
associated to that entity in that state.
It is used by the Entity class to then draw the corresponding asset.
State and StateInterface have been removed, since their functionality is
now deprecated due to these changes.